Signal a package-scoped message

Signal a package-scoped message

Usage

.pkg_inform(
  message,
  subclass,
  parent = NULL,
  call = rlang::caller_env(),
  message_env = rlang::caller_env(),
  ...
)

Arguments

message

(character) The message for the new message condition. Messages will be formatted with cli::cli_bullets().

subclass

(character) Class(es) to assign to the message. Will be prefixed by "{package}-message-".

parent

A parent condition, as you might create during a rlang::try_fetch(). See rlang::abort() for additional information.

call

(environment) The caller environment for error messages.

message_env

(environment) The execution environment to use to evaluate variables in error messages.

...

Additional parameters passed to cli::cli_inform() and on to rlang::inform().

Value

(NULL, invisibly) Called for message side effect.