Read a beekeeper config file

Read a YAML file (default name _beekeeper.yml) with configuration information about the API and the package. This file is generated by use_beekeeper(), or can be created manually.

Usage

read_config(pkg_dir = ".", config_filename = "_beekeeper.yml")

Arguments

pkg_dir

(character(1) or fs_path) The directory containing package files.

config_filename

(character(1) or fs_path) The path to a beekeeper yaml config file (relative to the package root).

Value

(list) Configuration information, with elements api_title, api_abbr, api_version, rapid_filename, and updated_on, plus security_data_filename when present.

Details

The configuration information must contain these fields:

api_title (character(1)): The title of the API, used in documentation and messages.

api_abbr (character(1)): An abbreviation for the API, used in function names and other identifiers.

api_version (character(1)): The version of the API.

rapid_filename (character(1)): The name of the file (relative to the package root) where the API definition is stored as an RDS file. By default, this is _beekeeper_rapid.rds, and is generated by use_beekeeper() based on an OpenAPI definition file.

updated_on (character(1)): A timestamp of when the API definition was last updated, in the format "YYYY-MM-DD HH:MM:SS" (UTC).

security_data_filename (character(1)): Optionally, the path (relative to the package root) where generated security metadata is stored.

See also

Examples

read_config(pkg_dir = fs::path_package("beekeeper", "guru"))
#> $api_title
#> [1] "APIs.guru"
#> 
#> $api_abbr
#> [1] "guru"
#> 
#> $api_version
#> [1] "2.2.0"
#> 
#> $rapid_filename
#> [1] "_beekeeper_rapid.rds"
#> 
#> $security_data_filename
#> [1] "_beekeeper_security.yml"
#> 
#> $updated_on
#> [1] "2026-05-12 07:57:02 UTC"
#>