Configure a package to use beekeeper

Create a configuration file for a package to use beekeeper. The configuration file tracks information that will be used for generation of other functions, and the timestamp when the configuration was last updated or used.

Usage

use_beekeeper(
  x,
  api_abbr,
  ...,
  config_file = "_beekeeper.yml",
  rapid_file = "_beekeeper_rapid.rds"
)

Arguments

x

An object to use to define the configuration. It must be translatable to a rapid::class_rapid() object by rapid::as_rapid(). Usually this will be a url pointing to an OpenAPI document, or a list generated by reading such a document.

api_abbr

A short (about 2-5 letter) abbreviation for the API, for use in function names and environment variables.

...

These dots are for future extensions and must be empty.

config_file

The path to which the configuration should be written.

rapid_file

The path to which the R API definition (rapid) object should be written.

Value

The path to the configuration file, invisibly. The config file is written as a side effect of this function. The rapid object is also written, and the path to that file is saved in the config file.