Configure a package to use beekeeper

Create a configuration file for a package to use beekeeper. The configuration file tracks information that will be used for generation of other functions, and the timestamp when the configuration was last updated or used.

Usage

use_beekeeper(
  x,
  api_abbr,
  ...,
  pkg_dir = ".",
  config_filename = "_beekeeper.yml",
  rapid_filename = "_beekeeper_rapid.rds"
)

Arguments

x

An object to use to define the configuration. It must be translatable to a rapid::class_rapid() object by rapid::as_rapid(). Usually this will be a url pointing to an OpenAPI document, or a list generated by reading such a document.

api_abbr

(character(1)) A short (about 2-5 letter) abbreviation for the API, for use in function names and environment variables.

...

These dots are for future extensions and must be empty.

pkg_dir

(character(1) or fs_path) The directory containing package files.

config_filename

(character(1) or fs_path) The path to a beekeeper yaml config file (relative to the package root).

rapid_filename

(character(1) or fs_path) The path to the R API definition (rapid) file (relative to the package root).

Value

(character(1), invisibly) The path to the configuration file. The config file is written as a side effect of this function. The rapid object is also written, and the path to that file (relative to pkg_dir) is saved in the config file.

Examples

# Set up an empty package.
pkg_dir <- unclass(fs::path_norm(withr::local_tempdir()))
usethis::create_package(pkg_dir, open = FALSE, check_name = FALSE)
#> ✔ Creating /tmp/RtmprMicgs/file52e469b5419/.
#> ✔ Setting active project to "/tmp/RtmprMicgs/file52e469b5419".
#> ✔ Creating R/.
#> ✔ Writing DESCRIPTION.
#> Package: file52e469b5419
#> Title: What the Package Does (One Line, Title Case)
#> Version: 0.0.0.9000
#> Authors@R (parsed):
#>     * First Last <first.last@example.com> [aut, cre]
#> Description: What the package does (one paragraph).
#> License: `use_mit_license()`, `use_gpl3_license()` or friends to
#>     pick a license
#> Encoding: UTF-8
#> Roxygen: list(markdown = TRUE)
#> RoxygenNote: 8.0.0
#> ✔ Writing NAMESPACE.
#> ✔ Setting active project to "<no active project>".
usethis::local_project(pkg_dir)
#> ✔ Setting active project to "/tmp/RtmprMicgs/file52e469b5419".
#> ✔ Setting active project to "<no active project>".

# Read an api definition. This could also simply be a URL to such a
# definition.
api_definition <- read_api_definition(fs::path_package("beekeeper", "guru"))

# Set up the package to use beekeeper with that definition.
use_beekeeper(api_definition, "guru")
#> ✔ Setting active project to "/__w/beekeeper/beekeeper".
#> ✔ Setting active project to "<no active project>".
#> ✔ Setting active project to "/__w/beekeeper/beekeeper".
#> ✔ Setting active project to "<no active project>".
read_config()
#> $api_title
#> [1] "APIs.guru"
#> 
#> $api_abbr
#> [1] "guru"
#> 
#> $api_version
#> [1] "2.2.0"
#> 
#> $rapid_filename
#> [1] "_beekeeper_rapid.rds"
#> 
#> $updated_on
#> [1] "2026-07-21 14:02:30 UTC"
#> 
#> $api_definition_origin
#> $api_definition_origin$url
#> [1] "https://api.apis.guru/v2/openapi.yaml"
#> 
#> $api_definition_origin$format
#> [1] "openapi"
#> 
#> $api_definition_origin$version
#> [1] "3.0"
#> 
#> 
all.equal(read_api_definition(), api_definition)
#> [1] TRUE

# Clean up.
withr::deferred_run()
#> No deferred expressions to run